+++ server/controllers/diaryController/GetSingleDiary.js
@@ -1,4 +1,4 @@
-import GetSingleDiaryEntryQuery from './getSingleEntryQuery';
+import GetSingleDiaryEntryQuery from './getSingleEntryDbQuery';
 import authenticateUser from '../../middlewares/authenticateUser';
 import authenticateToken from '../../middlewares/authenticateToken';
 import doesUserExist from '../../middlewares/doesUserExist';
