+++ server/controllers/diaryController/getSingleEntryDbQuery.js
